[#5] handle_proxy_request를 별도 함수로 분리
